fix: Modal path by media fields
