+++ packages/strapi-plugin-upload/admin/src/containers/App/index.js
@@ -1,55 +1,29 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
-import { LoadingIndicatorPage, request, useUserPermissions } from 'strapi-helper-plugin';
-import { isEmpty } from 'lodash';
+import { LoadingIndicatorPage, useUserPermissions } from 'strapi-helper-plugin';
 import pluginId from '../../pluginId';
 import pluginPermissions from '../../permissions';
 import { AppContext } from '../../contexts';
 
 import HomePage from '../HomePage';
-import ConfigContext from '../../contexts/ConfigContext';
-import { getRequestUrl } from '../../utils';
+import ConfigProvider from '../../providers/ConfigProvider';
 
 const App = () => {
   const state = useUserPermissions(pluginPermissions);
-  const [config, setConfig] = useState({});
-
-  const fetchConfig = async () => {
-    const requestURL = getRequestUrl('config');
-    try {
-      const data = await request(requestURL, {
-        method: 'GET',
-      });
-      setConfig(data);
-    } catch (err) {
-      console.error(err);
-      strapi.notification.toggle({
-        type: 'warning',
-        message: { id: 'notification.error' },
-      });
-      setConfig({ input: { types: ['*/*'] } });
-    }
-  };
-
-  useEffect(() => {
-    if (!state.isLoading) {
-      fetchConfig();
-    }
-  }, [state.isLoading]);
 
   // Show a loader while all permissions are being checked
-  if (state.isLoading || isEmpty(config)) {
+  if (state.isLoading) {
     return <LoadingIndicatorPage />;
   }
 
   if (state.allowedActions.canMain) {
     return (
       <AppContext.Provider value={state}>
-        <ConfigContext.Provider value={config}>
+        <ConfigProvider>
           <Switch>
             <Route path={`/plugins/${pluginId}`} component={HomePage} />
           </Switch>
-        </ConfigContext.Provider>
+        </ConfigProvider>
       </AppContext.Provider>
     );
   }
